Remove dependency on http form HttpEasy so that it can be more easily reused.
